Crate toylang

Toylang

A functional toy language using Polish Notation which compiles to Rust for fun and experimentation (Windows only for now)

Private functions are provided for information, the main info is here:

Github or the compiler_runner for general usage instructions

debug_window_derive for an extra debug windows app

integration_tests for examples

Hello, world!

Toylang code:

@ "Hello, world!"

Compiles to Rust code:

fn main() {
    println!("{}", "Hello, world!".to_string());
}

Modules

ast 🔒

AST containing adjacency list of Elements, output string, parents array

compiler_runner

Instructions for using toylang CLI

debug_window_derive

A windows app to help with debugging development of Toylang

errors 🔒

This is meant to be an exhaustive list of errors possible in the compiler.

file 🔒

Stores information about the input File, for use in the Compiler

formatting

Only contains get_formatted_argname_argtype_pairs

integration_tests

Examples for use with the Toylang CLI

parse 🔒

Main Parser functions

Structs

Compiler

The main struct that is used from start to finish of compilation

DebugErrorStack

DebugLinesOfChars

DebugLinesOfTokens

DebugLogs

Functions

left_pad 🔒

Debug helper to add correct spacing before numbers

num_digits 🔒

Debug helper for left_pad

rem_first_and_last 🔒

Type Definitions

ErrorStack 🔒

LinesOfChars 🔒

LinesOfTokens 🔒

Tokens