Module toylang::ast::elements

Elements are the individual nodes in the AST.

Each is made of an ElementInfo like Int(Value), and a Vec of indexes, which are references to child Elements elsewhere in the AST.

Modules

append

Functions to append each Element to the AST

Structs

ElementsVec

Enums

ElementInfo

Functions

_cut_and_paste_element_infos 🔒

Fake function #1, only useful for copy/pasting all the elementinfo types!

_cut_and_paste_elements 🔒

Fake function #2, only useful for copy/pasting all the option_element types!

get_arg_index_by_name

Get the index of the Arg based on its name

get_argnames_from_argtokens

Get argnames from Arg tokens, but also update Arg tokens returntypes at same time

get_argtypes_from_argtokens

Get a vec of types based on child refs, assuming they are Types or Parens (containing a Dyn Fn with types)

get_constant_by_name

Get the Constant based on its name

get_constant_index_by_name

Get the index of the Constant based on its name

get_element_by_name

Finds the original element referred to, e.g. when using a variable name

get_elementinfo_type

Get type of an ElementInfo

get_formatted_dyn_fn_type_sig

Get Type sig of Dyn Fn

get_function_index_by_name

Get the index of the Function based on its name

get_inbuilt_function_by_name

Get the inbuiltFn by its name

get_inbuilt_function_index_by_name

Get the index of the InbuiltFn based on its name

get_inbuilt_type_index_by_name

Get the index of the Type based on its name

get_infered_type_of_any_element

Get the inferred type of an element where possible or ignore if not

get_infered_type_of_arg_element

Get inferred type of Arg

get_infered_type_of_constant_element

Get inferred type of Constant

get_infered_type_of_constantref_element

Get inferred type of Constantref

get_infered_type_of_functioncall_element

Get inferred type of FnCall

get_infered_type_of_if_element

Get inferred type of If

get_infered_type_of_inbuiltfunctioncall_element

Get inferred type of InbuiltFnCall

get_last_element

Get last element in ast

get_returntype_from_argtokens

Get returntype from children assuming one is a Type

get_updated_elementinfo_with_infered_type

Get ElementInfo but with inferred type added

replace_element_child

Replace a child ref of an element with a ref to another child

replace_funcdefwip_with_funcdef

Replace the WIP FuncDef placeholder, with the final FuncDef

write_indent 🔒

write_subtree 🔒

Type Definitions

ArgModifier 🔒

ArgModifiers 🔒

ArgNames 🔒

ArgTypes 🔒

ElIndex

Element

ElementChildren

Elements

Format 🔒

From 🔒

Name 🔒

RefName 🔒

ReturnType 🔒

Scope 🔒

To 🔒

Value 🔒